Use `basename` on IO::Path object directly
It is no longer necessary to extract the path from an `IO` object in order
to request such information as the basename.  This now works directly on the
`IO::Path` which is returned when using the `IO` role on a path-like string.
This change corrects this issue and removes one source of errors in the test
suite.